Valencies of compound now checked
The method is complicated, will simplify in the future.
Also simplified try except by putting it in lowest_position() itself. (can be simplified more)
Added custom exception if valency is not satisfied
show_structure() a separate method to just see structure of compound
Not sure what to do by counting carbons and hydrogens, might remove in the future.