Take user input
Although this is meant to be a `module`, many people would prefer to
run the program directly. So, I have added the functionality to ask
the user for a valid chemical structure.

I have also added a `help` command to get usage information.

TODO: Maybe add more examples to the help() function?